+++ db/schema.rb
@@ -10,7 +10,7 @@
 #
 # It's strongly recommended that you check this file into your version control system.
 
-ActiveRecord::Schema.define(version: 2021_11_29_091202) do
+ActiveRecord::Schema.define(version: 2021_11_29_120610) do
 
   # These are extensions that must be enabled in order to support this database
   enable_extension "pgcrypto"
@@ -99,8 +99,13 @@ ActiveRecord::Schema.define(version: 2021_11_29_091202) do
     t.string "title"
     t.text "body"
     t.string "slug"
-    t.datetime "created_at", precision: 6, null: false
-    t.datetime "updated_at", precision: 6, null: false
+    t.datetime "created_at", default: -> { "CURRENT_TIMESTAMP" }, null: false
+    t.datetime "updated_at", default: -> { "CURRENT_TIMESTAMP" }, null: false
+    t.string "contentful_id"
+    t.text "sidebar"
+    t.string "breadcrumbs", default: [], array: true
+    t.index ["contentful_id"], name: "index_pages_on_contentful_id", unique: true
+    t.index ["slug"], name: "index_pages_on_slug", unique: true
   end
 
   create_table "radio_answers", id: :uuid, default: -> { "gen_random_uuid()" }, force: :cascade do |t|